Updates
Added a global timer that all events in the game use. Can be changed to
speed up/slow down the game.
Added a method for random calls. This is mostly for tidying up the code.

Added a method for colouring text as it gets sent to the games output.
Added some comments.
Added a check for combat so you can't return to town if you're fighting.

Changed updateGameText() so that it adds <br/> onto the end of every
string sent to it. This cuts down on repetition.
Changed a few variable names to make them easier to read.

Removed some variables that weren't necessary in monster, combat,
healthBarUpdate.

Probably some other stuff I forgot to!